Have method predicates as top level classes
